change : separate class from MergedModel.ts
